test: move authentication tests to LoginTest
The authentication tests were previously located in `AuthenticationTest.php`, which was a separate test file. However, these tests are specifically related to the login functionality provided by the `Login` Livewire component.

To improve the organization and maintainability of the test suite, the authentication tests have been moved to the `LoginTest.php` file, which already contains other tests for the `Login` component. The original `AuthenticationTest.php` file has been removed.

This change ensures that all tests related to the login functionality are grouped together in a single file, making it easier to understand and maintain the tests.
